refactor: 클래스 정리 완료
